Brought it into the 21st century (ish)
Powerline has changed and we gotta change with it. So, here's to being
up to date. I've made some changes to keep up with what Powerline
has done to make life a little easier.

This also means that the Lokaltog default fonts provided will now work
with stock powerline theme.